+++ nappgui/examples/guihello/splits.rs
@@ -3,7 +3,7 @@ use nappgui::prelude::*;
 static LOREM: &str = "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.";
 
 fn left_panel() -> Panel {
-    const N: usize = 32;
+    const N: u32 = 32;
     let panel = Panel::new_scroll(false, true);
     let layout = Layout::new(2, N);
     let rmargin = panel.scroll_width();
